chore: separate comunication into a component
