+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8">
-  <link rel="icon" href="/favicon.ico">
+  <link rel="icon" href="/src/assets/logo/icon-color.png" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link href="/src/assets/main.css" rel="stylesheet">
   <title>Vite App</title>
@@ -12,6 +12,25 @@
 <body>
   <div id="app"></div>
   <script type="module" src="/src/main.ts"></script>
+  <script>
+    const favicons = {
+      light: '/src/assets/logo/icon-light-trasp.png',
+      dark: '/src/assets/logo/icon-dark-trasp.png'
+    };
+
+    function getCurrentTheme() {
+      return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
+    }
+    function updateFavicon() {
+      const theme = getCurrentTheme();
+      const favicon = document.querySelector('link[rel="icon"]');
+      favicon.href = favicons[theme];
+    }
+
+    updateFavicon();
+
+    window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', updateFavicon);
+  </script>
 </body>
 
 </html>
