+++ src/views/HomeView.vue
@@ -1,16 +1,6 @@
 <script setup lang="ts">
-import { ref, computed, onMounted } from 'vue'
-
-interface Communication {
-  id: number | string
-  title: string
-  content: string
-  date: string
-  author: string
-  category: string
-  priority: string
-  cta?: { text: string; link: string } | null
-}
+import { ref } from 'vue'
+import ComunicationSection from '@/components/ComunicationSection.vue'
 
 interface Representative {
   id: number | string
@@ -21,10 +11,8 @@ interface Representative {
   year?: string
   curriculum?: string
 }
-import { RouterLink } from 'vue-router'
 import { Icon } from '@iconify/vue'
 
-const communications = ref<Communication[]>([])
 const representatives = ref<Representative[]>([
   {
     id: 1,
@@ -54,48 +42,7 @@ const representatives = ref<Representative[]>([
     curriculum: 'Triennale'
   }
 ])
-const selectedCategory = ref('all')
 
-// Categorie per il filtro
-const categories = [
-  { id: 'all', name: 'Tutte', icon: 'heroicons:clipboard-document-list' },
-  // { id: 'questionario', name: 'Questionari', icon: 'heroicons:document-text' },
-  { id: 'didattica', name: 'Didattica', icon: 'heroicons:academic-cap' },
-  { id: 'opportunita', name: 'Opportunità', icon: 'heroicons:rocket-launch' },
-  { id: 'eventi', name: 'Eventi', icon: 'heroicons:calendar-days' }
-]
-
-// Comunicazioni filtrate
-const filteredCommunications = computed(() => {
-  if (selectedCategory.value === 'all') {
-    return communications.value
-  }
-  return communications.value.filter(comm => comm.category === selectedCategory.value)
-})
-
-// Funzioni utili
-const formatDate = (dateString: string) => {
-  const date = new Date(dateString)
-  return date.toLocaleDateString('it-IT', {
-    day: 'numeric',
-    month: 'long',
-    year: 'numeric'
-  })
-}
-
-const getPriorityClass = (priority: string) => {
-  switch (priority) {
-    case 'high': return 'border-l-4 border-error bg-error/10'
-    case 'medium': return 'border-l-4 border-warning bg-warning/10'
-    case 'low': return 'border-l-4 border-info bg-info/10'
-    default: return 'border-l-4 border-base-300 bg-base-200/50'
-  }
-}
-
-const getCategoryIcon = (category: string) => {
-  const cat = categories.find(c => c.id === category)
-  return cat?.icon || 'heroicons:clipboard-document-list'
-}
 
 // Smooth scroll
 const scrollToSection = (sectionId: string) => {
@@ -105,28 +52,6 @@ const scrollToSection = (sectionId: string) => {
   }
 }
 
-// Load communications from generated JSON
-onMounted(async () => {
-  try {
-    const res = await fetch('/communications.json')
-    if (res.ok) {
-      const data = await res.json()
-      // Basic validation and fallback mapping
-      communications.value = Array.isArray(data) ? data.map((m: any) => ({
-        id: m.id,
-        title: m.title || (m.content ? String(m.content).slice(0, 80) + '…' : 'Senza titolo'),
-        content: m.content || '',
-        date: m.date || new Date().toISOString(),
-        author: m.author || 'Telegram',
-        category: m.category || 'didattica',
-        priority: m.priority || 'low',
-        cta: m.cta || null
-      })) : []
-    }
-  } catch (e) {
-    console.error('Errore caricamento communications.json', e)
-  }
-})
 </script>
 
 <template>
@@ -182,82 +107,7 @@ onMounted(async () => {
     </div>
   </section>
 
-  <!-- Communications Section -->
-  <section id="communications" class="py-20 bg-base-100">
-    <div class="container mx-auto px-4 max-w-6xl">
-      <!-- Section Header -->
-      <div class="text-center mb-16">
-        <h2 class="text-4xl md:text-5xl mb-4 text-base-content flex items-center justify-center gap-3">
-          <Icon icon="heroicons:megaphone" class="h-12 w-12" />
-          Comunicazioni
-        </h2>
-        <p class="text-xl text-base-content/70 max-w-2xl mx-auto">
-          Resta aggiornato su tutte le novità del corso di laurea
-        </p>
-      </div>
-
-      <!-- Category Filter -->
-      <div class="flex flex-wrap justify-center gap-3 mb-12">
-        <button v-for="category in categories" :key="category.id" @click="selectedCategory = category.id" :class="[
-          'btn btn-sm transition-all duration-200',
-          selectedCategory === category.id
-            ? 'btn-primary'
-            : 'btn-outline btn-primary hover:btn-primary'
-        ]">
-          <Icon :icon="category.icon" class="h-4 w-4 mr-2" />
-          {{ category.name }}
-        </button>
-      </div>
-
-      <!-- Communications Grid -->
-      <div class="grid gap-6 md:gap-8">
-        <div v-for="comm in filteredCommunications" :key="comm.id" :class="[
-          'card bg-base-100 shadow-lg hover:shadow-2xl transition-all duration-300 transform hover:-translate-y-1',
-          getPriorityClass(comm.priority)
-        ]">
-          <div class="card-body">
-            <!-- Header -->
-            <div class="flex flex-col sm:flex-row sm:items-start sm:justify-between gap-4 mb-4">
-              <div class="flex items-start gap-3">
-                <Icon :icon="getCategoryIcon(comm.category)" class="h-6 w-6 text-primary" />
-                <div>
-                  <div class="card-title text-lg md:text-xl mb-2 text-base-content" v-html="comm.title">
-                  </div>
-                  <div class="flex items-center gap-2 text-sm text-base-content/60">
-                    <Icon icon="heroicons:calendar-days" class="h-4 w-4" />
-                    {{ formatDate(comm.date) }}
-                  </div>
-                </div>
-              </div>
-
-              <div class="badge badge-primary badge-outline">
-                {{ comm.category }}
-              </div>
-            </div>
-
-            <!-- Content -->
-            <div class="text-base-content/80 leading-relaxed mb-4" v-html="comm.content"></div>
-
-
-            <!-- Footer -->
-            <div class="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4">
-              <div class="text-sm text-base-content/60 flex items-center gap-1">
-                <Icon icon="heroicons:pencil-square" class="h-4 w-4" />
-                {{ comm.author }}
-              </div>
-
-              <div v-if="comm.cta" class="flex justify-end">
-                <RouterLink :to="comm.cta.link" class="btn btn-primary btn-sm">
-                  {{ comm.cta.text }}
-                  <Icon icon="heroicons:arrow-right" class="h-4 w-4 ml-2" />
-                </RouterLink>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
+  <ComunicationSection></ComunicationSection>
 
   <!-- Representatives Section -->
   <section id="representatives" class="py-20 bg-base-200">
@@ -277,10 +127,10 @@ onMounted(async () => {
       <div class="grid md:grid-cols-2 lg:grid-cols-3 gap-8">
         <div v-for="rep in representatives" :key="rep.id"
           class="card bg-base-100 shadow-lg hover:shadow-2xl transition-all duration-300 transform hover:-translate-y-2">
-          <div class="card-body text-center">
+          <div class="card-body items-center">
             <!-- Avatar -->
             <div class="avatar placeholder mb-4">
-              <div class="bg-gradient-to-br from-[#0097b2] to-[#7ed957] text-white rounded-full w-20 h-20">
+              <div class="bg-gradient-to-br from-[#0097b2] to-[#7ed957] text-white rounded-full w-20 h-20 flex items-center justify-center">
                 <span class="text-2xl font-bold">{{rep.name.split(' ').map((n: string) => n[0]).join('')}}</span>
               </div>
             </div>
@@ -291,14 +141,14 @@ onMounted(async () => {
             <p class="text-base-content/60 text-xs mb-4">{{ rep.year }} • {{ rep.curriculum }}</p>
 
             <!-- Contact Buttons -->
-            <div class="flex gap-2 justify-center">
-              <a :href="`mailto:${rep.email}`" class="btn btn-sm btn-outline btn-primary" title="Invia Email">
-                <Icon icon="heroicons:envelope" class="h-4 w-4" />
+            <div class="flex gap-2 justify-center *:text-lg *:hover:text-xl">
+              <a :href="`mailto:${rep.email}`" class="btn btn-square btn-sm btn-ghost" title="Invia Email">
+                <Icon icon="heroicons-solid:mail" />
               </a>
 
               <a :href="`https://t.me/${rep.telegram.replace('@', '')}`" target="_blank"
-                class="btn btn-sm btn-outline btn-primary" title="Contatta su Telegram">
-                <Icon icon="logos:telegram" class="h-4 w-4" />
+                class="btn btn-square btn-sm btn-ghost" title="Contatta su Telegram">
+                <Icon icon="logos:telegram" />
               </a>
             </div>
           </div>
